requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
python-obd==0.7.1